+++ CHANGELOG.md
@@ -21,6 +21,7 @@
   - ＋ Checks that file with same name isn't uploaded before upload starts
   - ＋ Cancel a running upload if the upload-dialog is closed
 - ✓ Fixes user's can't log-out if forum is installed in a subdirectory
+- ✓ Fixes login redirect issues if forum is installed in a subdirecotry
 - Δ Improves performance of background task runner
 - Internal code changes:
   - Δ Increases phpstan static code analysis from level 3 to 4